I now have chapter skips working together with the slider bar.
Summary of fix:
   1) Inserted playlist items for each chapter.  Also inserted an item at the 
beginning of playlist for the root menu.  We can prevent user from seeing the root 
menu (via our media_player_dlg, anyway) by commenting out a single line of code.
   2) Added a function on_time_slider_button_move_event() that sets the VLC playback 
time to the current slider location then sleeps for 50 us (so as not to overload the 
system - this value may need to be tweaked).
   3) This new function is triggered by the slider bar's value_changed signal 
whenever the mouse has a hold of the slider bar.